+++ CHANGELOG.markdown
@@ -1,4 +1,8 @@
 # Changelog
+
+## social-wallet-api 0.3.0
+* Throttling behaviour is introduced for prevension from bruteforce attacks
+* New api calls for get-account were introduced
 	
 ## social-wallet-api 0.2.1
 Better explanation on how to use the lib with and without config. Implemented the list accounts call.